Use SvelteKit's goto() instead of location.href to avoid a server round-trip
This fixes the occasional race condition where IndexedDB hasn't saved yet the new transcript to disk
(despite firing the 'success' event stating it did so, due to default durability) while the browser
is navigating to a new url, causing a full page reload and a loss of Svelte state and IndexedDB transaction.
The user experience is an everloading details page, and a loss of the last new transcript.