Fix New Transcript not starting on Chrome for Android
Since micLevel and audioRecorder are disabled on Android, the user was never prompted
to grant microphone permission. Start speechRecognition first, to always get the prompt
then reset/start the rest of dependent stores and services.

Also turn speechRecognition into a readable store, like micLevel and audioRecorder, to
better control when the initialization code is run.